Add files via upload
Each time series dataset should be tested to determine if it exhibits stationarity, meaning 
that its statistical properties such as mean and variance are constant over time. Summary 
statistics are one of several methods used to assess the non-stationarity of a time series. 
This involves splitting the time series into two or more partitions and comparing the mean 
and variance of each group. If the means and variances differ significantly 
between the groups, the time series is likely non-stationary.